+++ app/services/stats/calculate_month.rb
@@ -63,11 +63,16 @@ class Stats::CalculateMonth
   def points
     return @points if defined?(@points)
 
+    country_name_column = if Point.connection.column_exists?(:points, :country_name_legacy)
+                            'COALESCE(points.country_name_legacy, points.country) AS country_name_legacy'
+                          else
+                            'points.country_name AS country_name_legacy'
+                          end
     @points = user
               .points
               .not_anomaly
               .where(timestamp: start_timestamp..end_timestamp)
-              .select(:id, :lonlat, :timestamp, :city, :country_id, :velocity)
+              .select(:id, :lonlat, :timestamp, :city, :country_id, :velocity, Arel.sql(country_name_column))
               .preload(:country)
               .order(timestamp: :asc)
   end